Simplifies the MALListDecoder interface & Refactors the HTTP stream reader to two classes
