+++ apis/api-mal/src/main/java/org/ccsds/moims/mo/mal/MALListDecoder.java
@@ -25,7 +25,7 @@ package org.ccsds.moims.mo.mal;
  *
  * Has non-standard extension size() operation.
  */
-public interface MALListDecoder extends MALDecoder {
+public interface MALListDecoder {
 
     /**
      * Determines if there are any more elements in the lists to decode.
