+++ bindings/transport-file/src/main/java/esa/mo/mal/encoder/line/LineDecoder.java
@@ -743,7 +743,7 @@ public class LineDecoder implements MALDecoder {
                 list.add(Attribute.attribute2JavaType(att));
             } else {
                 // Normal Case
-                list.add(element.decode(listDecoder));
+                list.add(element.decode(this));
             }
         }
         return list;
